# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SH3 Backend API",
        "description": "API de Atendimentos feita por Hugo Galli para SH3",
        "contact": {
            "name": "Hugo Galli",
            "email": "hugo.galli.r@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/areas": {
            "get": {
                "tags": [
                    "Áreas"
                ],
                "summary": "Lista de áreas",
                "description": "Retorna a lista de áreas",
                "operationId": "f16518e18a005a2f966b63c0e1745342",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "areas": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Area"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "jwt_token": []
                    }
                ]
            }
        },
        "/api/area": {
            "post": {
                "tags": [
                    "Áreas"
                ],
                "summary": "Criar uma nova área",
                "description": "Cria uma nova área",
                "operationId": "e0f26c363ff9ac10e23b932711798e65",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Título da área",
                                        "type": "string",
                                        "example": "Nova Área"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Área criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Área criada com sucesso"
                                        },
                                        "area": {
                                            "$ref": "#/components/schemas/Area"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "jwt_token": []
                    }
                ]
            }
        },
        "/api/area/{id}": {
            "get": {
                "tags": [
                    "Áreas"
                ],
                "summary": "Detalhes de uma área específica",
                "description": "Retorna os detalhes de uma área específica",
                "operationId": "cbf3e9187032439d3e6c110d1061a3ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da área",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "area": {
                                            "$ref": "#/components/schemas/Area"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Área não encontrada"
                    }
                },
                "security": [
                    {
                        "jwt_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Áreas"
                ],
                "summary": "Atualizar uma área existente",
                "description": "Atualiza uma área existente",
                "operationId": "56bfb71778faaf297d418afd709cc9bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da área",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Novo título da área",
                                        "type": "string",
                                        "example": "Novo Título"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Área atualizada com sucesso"
                                        },
                                        "area": {
                                            "$ref": "#/components/schemas/Area"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Área não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "jwt_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Áreas"
                ],
                "summary": "Excluir uma área",
                "description": "Exclui uma área",
                "operationId": "2e2ec2b139a849792cc2fb51cf460d0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da área",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Área excluída com sucesso"
                                        },
                                        "area": {
                                            "$ref": "#/components/schemas/Area"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Área não encontrada"
                    }
                },
                "security": [
                    {
                        "jwt_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Area": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Area A"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "jwt_token": {
                "type": "apiKey",
                "scheme": "bearer",
                "description": "Insira valores no formato (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}